recovery_backend: raise some failures during build if we know it won't work
isolinux is a Recommends: intentionally so that this can be included into
public Ubuntu images without having apt get upset, but it is needed to build
recovery partitions.

If the path for the EFI partition included in the image changes again raise
errors to make debugging easier in the future.